pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "agentic-project-kit"
version = "0.1.0"
description = "Generate GitHub-ready project templates for human-AI software development workflows."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Volker Fickert" }
]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
  "pydantic>=2.7"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6"
]

[project.scripts]
agentic-kit = "agentic_project_kit.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
